GCP/2.Excercises/Marco/UaaS/requirements.txt
flask == 2.3.3
requests
flask_restful
WTForms
google-cloud-pubsub
google-cloud-firestore
numpy
gunicorn